auth backend working, tests complete, frontend getting empty or no responses for some reason. CORS?

+++ tests/backendAuthTests.sh
@@ -4,37 +4,44 @@
 mongoUse='use easytrage'
 mongoClear='db.dropDatabase()'
 mongoExit='exit'
-{
+
 echo ""
 echo "========== Backend Authentication Tests =========="
+echo ""
 echo "WARNING: TEST SCRIPT WILL CLEAR MONGO-DB"
 echo ""
+read -p "Continue? (y/n)" CONT
 echo ""
-# (
-# docker exec -it mongo /bin/bash <<'EOF'
-# > mongo
-# > use easytrage
-# > db.dropDatabase()
-# > exit
-# EOF
-# )
-(
+if [ "$CONT" = "y" ]; then
+    (
+    echo ""
+
+    echo "===== clearing database ====="
+    echo ""
     docker exec -it mongo /bin/bash | mongo <<'EOF'
     use easytrage
     db.dropDatabase()
     exit
-    exit
 EOF
-)
-echo "database cleared"
-echo ""
-echo "Output of backend /register call with email 'vmelkote@ucsc.edu' and password 'password' : "
-echo ""
-curl -d '{"email":"vmelkote@ucsc.edu", "password":"password"}' -H "Content-Type: application/json" -X POST http://localhost:3001/api/register ;
-echo ""
-echo ""
-echo ""
-echo "Output of backend /login call with email 'vmelkote@ucsc.edu' and password 'password' : "
-curl -d '{"email":"vmelkote@ucsc.edu", "password":"password"}' -H "Content-Type: application/json" -X POST http://localhost:3001/api/login ;
-echo ""
-} > backendAuthTestsOutput.txt
+    <<< exit
+    echo ""
+    echo "===== database cleared ====="
+    echo ""
+    echo ""
+
+    echo "======== test output ========"
+    echo ""
+    echo "Output of backend /register call with email 'vmelkote@ucsc.edu' and password 'password' : "
+    echo ""
+    curl -d '{"email":"vmelkote@ucsc.edu", "password":"password"}' -H "Content-Type: application/json" -X POST http://localhost:3001/api/register ;
+    echo ""
+    echo ""
+    echo ""
+    echo "Output of backend /login call with email 'vmelkote@ucsc.edu' and password 'password' : "
+    echo ""
+    curl -d '{"email":"vmelkote@ucsc.edu", "password":"password"}' -H "Content-Type: application/json" -X POST http://localhost:3001/api/login ;
+    echo ""
+    echo ""
+    echo "======== end of tests ========"
+    ) > backendAuthTestsOutput.txt
+fi
